update: 사용자 인증 엔드포인트 위치 및 주소 변경
